requirements.txt
apache-airflow-providers-postgres
apache-airflow-providers-amazon
dbt-postgres
pandas
scikit-learn
minio
psycopg2-binary